pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dacapo-monai"
version = "1.0.0"
description = "MONAI transforms integration for DaCapo toolbox"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Marwan Zouinkhi", email = "zouinkhi.marwan@gmail.com"}
]
keywords = [
    "deep learning",
    "medical imaging",
    "computer vision",
    "pytorch",
    "monai",
    "dacapo",
    "transforms",
    "data augmentation"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "torch>=1.12.0",
    "numpy>=1.21.0",
    "dacapo-toolbox>=1.0.0",
    "monai>=1.0.0",
    "funlib-geometry>=0.2.0",
    "funlib-persistence>=0.6.0",
    "gunpowder>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "pytest-mock>=3.10.0",
    "black>=22.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "safety>=2.3.0",
    "bandit>=1.7.0",
]
docs = [
    "sphinx>=6.0.0",
    "sphinx-design>=0.4.0",
    "furo>=2023.3.27",
    "myst-parser>=1.0.0",
    "sphinx-copybutton>=0.5.0",
    "sphinxcontrib-mermaid>=0.8.0",
    "sphinx-autobuild>=2021.3.14",
    "sphinx-autodoc-typehints>=1.23.0",
]
examples = [
    "matplotlib>=3.5.0",
    "jupyter>=1.0.0",
    "tqdm>=4.62.0",
]
all = [
    "dacapo-monai[dev,docs,examples]",
]

[project.urls]
Homepage = "https://github.com/yourusername/dacapo-monai"
Repository = "https://github.com/yourusername/dacapo-monai"
Documentation = "https://dacapo-monai.readthedocs.io"
"Bug Tracker" = "https://github.com/yourusername/dacapo-monai/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"dacapo_monai" = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --cov=dacapo_monai --cov-report=term-missing"